Row-Col and Bayes approach to breeding
Scripts for Row-Col and Bayesian approach seeking to improve the predictive capacity and selection of passion fruit paper

André Oliveira Souza 1, Alexandre Pio Viana2, Fabyano Fonseca e Silva3, Camila Ferreira Azevedo4, Flavia Alves da Silva2*, Fernando Higino Lima e Silva5

1Federal Institute of Espírito Santo, Alegre, Espírito Santo, Brazil.
2Northern Fluminense State University Darcy Ribeiro (UENF), Plant Breeding Laboratory (LMGV), Av. Alberto Lamego, 2000, 28013-602, Campos dos Goytacazes - RJ, Brazil.
3Federal University of Viçosa, Department of Zootechnics, Av. Peter Henry Rolfs, 36570-000, Viçosa - MG, Brazil.
4Federal University of Viçosa, Department of Statistics, Av. Peter Henry Rolfs, 36570-000, Viçosa - MG, Brazil.
5Federal Institute of Education, Science and Technology from Goiás, South Highway Goiana Km 01, 75901-970, Rio Verde - GO, Brazil.
*Corresponding author e-mail: flavia_uems@hotmail.com 
ABSTRACT: Genetic improvement methods of semi-perennial species such as passion fruit often involve large areas, unbalanced data, and lack of observations. Some strategies can be used to work around these problems. In this work, different models and approaches were tested to improve the precision of estimates of genetic evaluation models for several characteristics of the passion fruit. A randomized block design (RBD) model was compared with a posteriori correction, adding two factors to the model (post-hoc blocking Row-Col). These models were also combined with the frequentist and Bayesian approaches, to see which combination yields the most accurate results. These approaches are part of a strategic plan in a perennial plant breeding program, in order to select promising genitors of passion to compose the next selection cycle. For Bayesian, we tested two prioris, defining different values for the parameters of the distributions of the variances of the effects of the model. A cross-validation test was also performed, both for choosing a priori values and comparing the frequentist and Bayesian approaches using the root mean square error (RMSE) and the correlation between the predicted and observed values, called Predictive capacity of the model (PC). The model that had the post-hoc blocking Row-Col design, captured the spatial variability for productivity and number of fruits, having a direct impact on the experimental precision. Both approaches applied to the models showed a similar performance, with predictive capacity and selective efficiency leading to the selection of the same individuals.